Handle scientific notation and unsigned.
M	include/vara/Feature/Constraint.h
M	include/vara/Feature/ConstraintParser.h
M	include/vara/Feature/Feature.h
M	include/vara/Feature/FeatureModelParser.h
M	lib/Feature/Feature.cpp
M	lib/Feature/FeatureModelParser.cpp
M	lib/Feature/FeatureModelWriter.cpp
M	unittests/Feature/ConstraintParser.cpp
M	unittests/Feature/FeatureModelBuilder.cpp
M	unittests/Feature/FeatureModelParser.cpp
M	unittests/Feature/FeatureModelTransaction.cpp
M	unittests/Feature/FeatureTreeNode.cpp
M	unittests/Feature/NumericFeature.cpp
M	unittests/Feature/Relationship.cpp
M	unittests/resources/CMakeLists.txt
A	unittests/resources/xml/test_numbers.xml

+++ include/vara/Feature/Constraint.h
@@ -570,10 +570,10 @@ public:
 
 class PrimaryIntegerConstraint : public PrimaryConstraint {
 public:
-  PrimaryIntegerConstraint(int Value)
+  PrimaryIntegerConstraint(long Value)
       : PrimaryConstraint(ConstraintKind::CK_INTEGER), Value(Value) {}
 
-  [[nodiscard]] int getValue() const { return Value; }
+  [[nodiscard]] long getValue() const { return Value; }
 
   std::unique_ptr<Constraint> clone() override {
     return std::make_unique<PrimaryIntegerConstraint>(this->getValue());
@@ -590,7 +590,7 @@ public:
   void accept(ConstraintVisitor &V) override;
 
 private:
-  int Value;
+  long Value;
 };
 
 class Feature;
